=fix calculation of kk on 32 bit computers

+++ src/constants.jl
@@ -2,7 +2,7 @@
 # Note: this code more or less comes directly from https://en.wikipedia.org/wiki/MD5
 # I believe this is fair use. and does not have lisence implications
 
-const kk =  floor.(UInt32, 2^32 * abs.(sin.(1:64)))
+const kk =  floor.(UInt32, Int64(2)^32 * abs.(sin.(1:64)))
 
 const ss = UInt64[
 7, 12, 17, 22,  7, 12, 17, 22,  7, 12, 17, 22,  7, 12, 17, 22,
